+++ src/mako/throttle/context/ContextInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\throttle\context;
+
+/**
+ * Context interface.
+ */
+interface ContextInterface
+{
+	/**
+	 * Returns the context identifier.
+	 */
+	public function getIdentifier(): string;
+}