+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 	- `Statistics::sampleStandardDeviation()`
 	- `Statistics::populationStandardDeviation()`
 * Added "update and return" functionality to the query builder. The feature is currently supported by the `Firebird`, `PostgreSQL`, `SQLite` and `SQL Server` compilers.
+* HTTP exceptions can now provide response headers.
+* Added a rate limiter library.
+* Added a rate limiter HTTP middleware.
 
 #### Changes